re-re-factor?
